% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_error.R
\name{odin_error_explain}
\alias{odin_error_explain}
\title{Explain an odin error}
\usage{
odin_error_explain(code, how = "pretty")
}
\arguments{
\item{code}{The error code, as a string, in the form \code{Exxxx} (a
capital "E" followed by four numbers)}

\item{how}{How to explain the error. Options are \code{pretty} (render
pretty text in the console), \code{plain} (display plain text in the
console) and \code{link} (browse to the online help).}
}
\value{
Nothing, this is called for its side effect only
}
\description{
Explain error codes produced by odin.  When odin fails to parse
your code (e.g., via \code{\link[=odin]{odin()}} or \code{\link[=odin_validate]{odin_validate()}}) it will return
an error with a code.  You can use \code{odin_error_explain} to get
more information on that code.  By default we will print an
explanation to the screen, but you can control this behaviour via
the \code{how} argument.  All error codes can be found in
\code{vignette("errors")}.
}
\examples{
odin_error_explain("E1006")
}
