% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate.R
\name{odin_migrate}
\alias{odin_migrate}
\title{Migrate odin code}
\usage{
odin_migrate(path, dest)
}
\arguments{
\item{path}{Path of the odin code to read}

\item{dest}{Path of the destination code.  It can be the same as
\code{dest}, in which case the file will be overwritten.}
}
\value{
Nothing; called for side effects only
}
\description{
Migrate odin code.  This function takes a path to existing odin
code and writes out migrated code to a new file.  It is possible
that no code will be migrated, in which case the written contents
will be identical to those read.
}
\examples{
\dontshow{
path <- tempfile(fileext = ".R")
writeLines(
  c("initial(x) <- 0",
    "deriv(x) <- x * r",
    "r <- user()"),
  path)
}
# A file 'path' contains odin code using old features:
writeLines(readLines(path))

# Migrate this file in place (by overwriting)
odin_migrate(path, path)

writeLines(readLines(path))
}
