% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odin.R
\name{odin_show}
\alias{odin_show}
\title{Show generated odin code}
\usage{
odin_show(expr, input_type = NULL, compatibility = "warning")
}
\arguments{
\item{expr}{Odin code as the path to a file (a string), a
character vector of code, or as an expression (typically within
braces \code{{}}).}

\item{input_type}{An optional string describing the type of input
for \code{expr} - must be one of \code{file}, \code{text} or \code{expression}.  If
given, this skips the type detection logic and odin will throw
an error if the wrong type of input is given.  Using this may be
beneficial in programmatic environments.}

\item{compatibility}{Compatibility mode to use.  Valid options are
"warning", which updates code that can be fixed, with warnings,
and "error", which will error.  The option "silent" will
silently rewrite code, but this is not recommended for general
use as eventually the compatibility mode will be removed (this
option is primarily intended for comparing output of odin1 and
odin2 models against old code).}
}
\value{
A character vector, with class \code{odin_code} that has a
pretty-print method defined.
}
\description{
Show generated code from compiling an odin model.
}
