% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odin.R
\name{odin}
\alias{odin}
\title{Compile an odin model}
\usage{
odin(
  expr,
  input_type = NULL,
  quiet = NULL,
  workdir = NULL,
  debug = NULL,
  skip_cache = FALSE,
  compatibility = NULL
)
}
\arguments{
\item{expr}{Odin code as the path to a file (a string), a
character vector of code, or as an expression (typically within
braces \code{{}}).}

\item{input_type}{An optional string describing the type of input
for \code{expr} - must be one of \code{file}, \code{text} or \code{expression}.  If
given, this skips the type detection logic and odin will throw
an error if the wrong type of input is given.  Using this may be
beneficial in programmatic environments.}

\item{quiet}{Logical, indicating if compilation messages from
\code{pkgbuild} should be displayed.  Error messages will be
displayed on compilation failure regardless of the value used.
If \code{NULL} is given, then we take the value from \code{DUST_QUIET} if set, or
\code{FALSE} otherwise.}

\item{workdir}{Optional working directory to use.  If \code{NULL}, the
behaviour depends on the existence of the environment variable
\code{DUST_WORKDIR_ROOT}.  If it is unset we use a session-specific
temporary directory (generated by \code{\link[=tempfile]{tempfile()}}).  If
\code{DUST_WORKDIR_ROOT} is set, then we use a stable generated
filename within this directory, which allows different sessions
to effectively share a cache.  If you pass a directory name here
as a string, then we will use that directory to write all code,
which allows you to inspect the generated code.  See
\code{vignette("details")} for more information.}

\item{debug}{Passed to \link[pkgbuild:compile_dll]{pkgbuild::compile_dll}, this will build a
debug library. If \code{NULL} is given, then we take the value from
\code{DUST_DEBUG} if set, or \code{FALSE} otherwise.}

\item{skip_cache}{Logical, indicating if the cache of previously
compiled systems should be skipped. If \code{TRUE} then your system will
not be looked for in the cache, nor will it be added to the
cache after compilation.}

\item{compatibility}{Compatibility mode to use.  Valid options are
"warning", which updates code that can be fixed, with warnings,
and "error", which will error.  The option "silent" will
silently rewrite code, but this is not recommended for general
use as eventually the compatibility mode will be removed (this
option is primarily intended for comparing output of odin1 and
odin2 models against old code).  The default, \code{NULL}, currently
corresponds to \code{warning}.}
}
\value{
A \code{dust_system_generator} object, suitable for using with
dust functions (starting from \link[dust2:dust_system_create]{dust2::dust_system_create})
}
\description{
Compile an odin model, yielding a \code{dust_system_generator} object.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# A random walk:
gen <- odin({
  initial(x) <- 0
  update(x) <- Normal(x, 1)
})

sys <- dust2::dust_system_create(gen, list(), n_particles = 10)
y <- dust2::dust_system_simulate(sys, 0:100)
matplot(t(y[1, , ]), type = "l", lty = 1, xlab = "Time", ylab = "Value")
\dontshow{\}) # examplesIf}
}
