% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odin.R
\name{odin}
\alias{odin}
\title{Compile an odin model}
\usage{
odin(
  expr,
  input_type = NULL,
  quiet = FALSE,
  workdir = NULL,
  debug = FALSE,
  skip_cache = FALSE
)
}
\arguments{
\item{expr}{Odin code as the path to a file (a string), a
character vector of code, or as an expression (typically within
braces \code{{}}).}

\item{input_type}{An optional string describing the type of input
for \code{expr} - must be one of \code{file}, \code{text} or \code{expression}.  If
given, this skips the type detection logic and odin will throw
an error if the wrong type of input is given.  Using this may be
beneficial in programmatic environments.}

\item{quiet}{Logical, indicating if compilation messages from
\code{pkgbuild} should be displayed.  Error messages will be
displayed on compilation failure regardless of the value used.}

\item{workdir}{Optional working directory to use.  If \code{NULL}, we
work in the session-specific temporary directory.  By using a
different directory of your choosing you can see the generated
code.}

\item{debug}{Passed to \link[pkgbuild:compile_dll]{pkgbuild::compile_dll}, this will build a
debug library.}

\item{skip_cache}{Logical, indicating if the cache of previously
compiled systems should be skipped. If \code{TRUE} then your system will
not be looked for in the cache, nor will it be added to the
cache after compilation.}
}
\value{
A \code{dust_system_generator} object, suitable for using with
dust functions (starting from \link[dust2:dust_system_create]{dust2::dust_system_create})
}
\description{
Compile an odin model, yielding a \code{dust_system_generator} object.
}
