% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_error.R
\name{odin_error_explain}
\alias{odin_error_explain}
\title{Explain odin error}
\usage{
odin_error_explain(code, how = "pretty")
}
\arguments{
\item{code}{The error code, as a string, in the form \code{Exxxx} (a
capital "E" followed by four numbers)}

\item{how}{How to explain the error. Options are \code{pretty} (render
pretty text in the console), \code{plain} (display plain text in the
console) and \code{link} (browse to the online help).}
}
\value{
Nothing, this is called for its side effect only
}
\description{
Explain error codes produced by odin.  This is a work in progress,
and we would like feedback on what is useful as we improve it.
The idea is that if you see an error you can link through to get
more information on what it means and how to resolve it.  The
current implementation of this will send you to the rendered
vignettes, but in future we will arrange for offline rendering
too.
}
